+++ ninja-core/src/main/java/ninja/utils/SwissKnife.java
@@ -138,7 +138,6 @@ public class SwissKnife {
         toAsNonPrimitiveType = Primitives.wrap(to);
 
         if (toAsNonPrimitiveType.isAssignableFrom(from.getClass())) {
-            System.out.println("assignable from!!!");
             return toAsNonPrimitiveType.cast(from);
         }